week_02/requirements.txt
zenml[server]>=0.50.0
requests>=2.25.0
pandas>=1.3.0
sentence-transformers>=2.2.0
numpy>=1.21.0
qdrant-client>=1.6.0
pymongo>=4.5.0
python-dotenv>=1.0.0